obj-detect/requirements.txt
absl-py==0.2.2
apturl==0.5.2
asn1crypto==0.24.0
astor==0.6.2
astroid==1.6.5
awscli==1.15.31
backcall==0.1.0
bleach==1.5.0
botocore==1.10.31
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.28.3
decorator==4.3.0
defer==1.0.6
distro-info==0.18
docutils==0.14
entrypoints==0.2.3
first==2.0.1
future==0.16.0
gast==0.2.0
gevent==1.3.3
greenlet==0.4.13
grpcio==1.12.0
html5lib==0.9999999
httplib2==0.9.2
idna==2.6
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
msgpack-python==0.5.6
mysql-connector==2.1.6
mysql-connector-python==2.1.6
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.14.3
oauth==1.0.1
olefile==0.45.1
pandocfilters==1.4.2
parso==0.2.1
pbr==4.0.3
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.2.0
pip-tools==2.0.2
pkg-resources==0.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pyasn1==0.4.3
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
pylint==1.9.2
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.2
python-dateutil==2.7.3
python-debian==0.1.32
python-magic==0.4.16
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
rsa==3.4.2
s3cmd==2.0.1
s3transfer==0.1.13
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
system-service==0.3
systemd-python==234
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
testresources==2.0.1
tornado==5.0.2
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
xkit==0.0.0
zerorpc==0.6.1
zope.interface==4.3.2
